requirements.txt
suds-community
argparse
ssl
prettytable